+++ subpixal/cc.py
@@ -19,7 +19,7 @@ __all__ = ['find_displacement']
 
 
 def find_displacement(ref_image, image00, image10, image01, image11,
-                      full_output=False):
+                      cc_type='NCC', full_output=False):
     """
     Find subpixel displacements between one reference cutout and a set of
     four "dithered" cutouts. This is achieved by finding peak position in a
@@ -47,6 +47,19 @@ def find_displacement(ref_image, image00, image10, image01, image11,
         "Same" image as ``image00`` but sampled at a 1/2 pixel displacement
         along both X-axis and Y-axis. It must have same shape as ``ref_image``.
 
+    cc_type : {'CC', 'NCC', 'ZNCC'}, optional
+        Cross-correlation algorithm to be used. ``'CC'`` indicates the
+        "standard" cross-correlation algorithm. ``'NCC'`` refers to the
+        normalized cross-correlation and ``'ZNCC'`` refers to the
+        zero-normalized cross-correlation, see, e.g.,
+        `Terminology in image processing \
+        <https://en.wikipedia.org/wiki/Cross-correlation#\
+        Terminology_in_image_processing>`_.
+
+    full_output : bool, optional
+        Return displacements only (``full_output=False``) or also interlaced
+        cross-correlation image and direct (non-interlaced) cross-correlation images
+
     Returns
     -------
 
@@ -62,8 +75,14 @@ def find_displacement(ref_image, image00, image10, image01, image11,
         Interlaced ("supersampled") cross-correlation image. Returned only when
         ``full_output`` is `True`.
 
+    ccs : list of numpy.ndarray, Optional
+        List of cross-correlation images between ``ref_image`` and ``image??``.
+        Returned only when ``full_output`` is `True`.
+
     """
-    icc, ccs = _build_icc_image(ref_image, image00, image10, image01, image11)
+    icc, ccs = _build_icc_image(
+        ref_image, image00, image10, image01, image11, cc_type
+    )
     xm, ym = find_peak(icc, peak_fit_box=5, peak_search_box='all')
 
     # find center of the auto-correlation peak when using signal.fftconvolve:
@@ -76,7 +95,7 @@ def find_displacement(ref_image, image00, image10, image01, image11,
     return (dx, dy, icc, ccs) if full_output else (dx, dy)
 
 
-def _build_icc_image(ref, im00, im10, im01, im11):
+def _build_icc_image(ref, im00, im10, im01, im11, cc_type='NCC'):
     """ Build interlaced ("oversampled") cross-correlation image. """
     # Check that all images have the same size. This is not really required,
     # but it simplifies things and, moreover, the rest of the code is designed
@@ -84,6 +103,10 @@ def _build_icc_image(ref, im00, im10, im01, im11):
     if not np.all(np.equal(ref.shape,
                            [im00.shape, im10.shape, im01.shape, im11.shape])):
         raise ValueError("All cutouts must have same shape.")
+
+    cc_type = cc_type.upper()
+    if cc_type in ['ZNCC', 'NCC']:
+        _normalize(ref, [im00, im10, im01, im11], cc_type == 'ZNCC')
 
     # cross-correlate images of different shifts
     cc00 = signal.fftconvolve(ref, im00[::-1, ::-1], mode='same')
@@ -101,3 +124,26 @@ def _build_icc_image(ref, im00, im10, im01, im11):
     icc[1::2, 1::2] = cc11[::-1, ::-1]
 
     return icc, (cc00, cc10, cc01, cc11)
+
+
+def _normalize(ref, images, zero=False):
+    data = []
+    masks = []
+    for im in images:
+        mask = im != 0
+        masks.append(mask)
+        data.append(im[mask])
+
+    data = np.hstack(data)
+
+    mean = np.mean(data) if zero else 0.0
+    std = np.std(data)
+
+    for im, m in zip(images, masks):
+        im[m] -= mean
+        im[m] /= std
+
+    mask = np.sum(masks, axis=0, dtype=np.bool)
+    if zero:
+        ref -= np.mean(ref[mask])
+    ref /= np.std(ref[mask])