+++ CHANGELOG.rst
@@ -12,6 +12,9 @@ Release Notes
 0.0.6.dev (26-September-2019)
 =============================
 
+- Added support for zero-normalized cross-correlation (ZNCC) and normalized
+  cross-correlation (NCC) algorithms. [#42]
+
 - Allow alignment code to run with without cosmic ray-cleaned images. [#41]
 
 - Reliability enhancement in handling cases when sky computation is